requirements.txt
fastapi==0.104.1
motor==3.3.2
pydantic==2.5.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
pytest==7.4.3
httpx==0.25.2
uvicorn[standard]==0.24.0

